+++ changelog.md
@@ -20,6 +20,7 @@ See the full changelog at https://github.com/iamkaf/amber
 - Added a KeybindHelper to register keybinds.
 - Added events to render to the HUD.
 - Added the `/amber doctor` command to diagnose issues with the mod.
+- Added client tick events.
 
 ### Changed